pyproject.toml
[project]
name = "simextract"
version = "0.1.0"
description = "CSV data vectorization and ChromaDB storage system"
requires-python = ">=3.9"
dependencies = [
    "chromadb>=0.4.22",
    "pandas>=2.0.0",
    "requests>=2.31.0",
    "tqdm>=4.66.0",
    "pyyaml>=6.0.1",
    "colorlog>=6.8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
